+++ NEWS.md
@@ -1,6 +1,9 @@
 # lasR 0.2.1
 
 - Fix: `callback()` properly handles errors from the injected function
+- New: handy functions `tempxyz()` to generate temp files with extension `.xyz`.
+- New: `rasterize()` is now parallelized with internal metrics.
+- New: `rasterize()` gained a progress bar with internal metrics.
 
 # lasR 0.2.0 (2024-03-01)